requirements.txt
# --- UI ---
streamlit==1.41.1

# --- PDF parsing ---
pymupdf==1.24.14

# --- RAG / vector search ---
faiss-cpu>=1.9.0.post1

# --- Embeddings ---
sentence-transformers==3.3.1

# --- LLM / summarization / translation ---
transformers==4.47.1
torch==2.5.1
sentencepiece==0.2.0
sacremoses==0.1.1

# --- Utils ---
numpy==1.26.4
tqdm==4.67.1
scikit-learn==1.5.2
